+++ R/bind_rows.R
@@ -5,7 +5,7 @@
 NULL
 
 # list length>=1 no null entries
-r_replyr_bind_rows <- function(lst,eagerCompute) {
+r_replyr_bind_rows <- function(lst, eagerCompute) {
   n <- length(lst)
   if(n<=1) {
     res <- lst[[1]]
@@ -22,7 +22,12 @@ r_replyr_bind_rows <- function(lst,eagerCompute) {
   # ideas from https://github.com/rstudio/sparklyr/issues/76
   # would like to use union_all, but seems to have problems with Spark 2.0.0
   # (spread example from basicChecksSpark200.Rmd)
-  if(length(intersect(c('spark_connection','src_spark'),replyr_dataServiceName(left)))>0) {
+  # make sure columns are in same order (dplyr::union over Spark doesn't like disorder)
+  # see: https://github.com/WinVector/replyr/blob/master/issues/UnionIssue.md
+  colnams <- intersect(colnames(left), colnames(right))
+  left <- dplyr::select_(left, .dots=colnams)
+  right <- dplyr::select_(right, .dots=colnams)
+  if(length(intersect(c('spark_connection','src_spark'), replyr_dataServiceName(left)))>0) {
     res <- dplyr::union(left,right)
   } else {
     res <- dplyr::union_all(left,right)
@@ -46,7 +51,7 @@ r_replyr_bind_rows <- function(lst,eagerCompute) {
 #' replyr_bind_rows(list(d,d,d))
 #'
 #' @export
-replyr_bind_rows <- function(lst,eagerCompute=FALSE) {
+replyr_bind_rows <- function(lst, eagerCompute= TRUE) {
   if(("NULL" %in% class(lst))||(length(lst)<=0)) {
     return(NULL)
   }
@@ -56,5 +61,5 @@ replyr_bind_rows <- function(lst,eagerCompute=FALSE) {
     return(NULL)
   }
   names(list) <- NULL
-  r_replyr_bind_rows(lst,eagerCompute)
+  r_replyr_bind_rows(lst, eagerCompute)
 }