+++ R/copyToFrom.R
@@ -2,9 +2,60 @@
 # Contributed by John Mount jmount@win-vector.com , ownership assigned to Win-Vector LLC.
 # Win-Vector LLC currently distributes this code without intellectual property indemnification, warranty, claim of fitness of purpose, or any other guarantee under a GPL3 license.
 
-#' @importFrom dplyr collect copy_to
+#' @importFrom dplyr collect copy_to db_drop_table
 NULL
 
+
+#' Drop a table from a source
+#'
+#' @param dest remote data source
+#' @param name name of table to drop
+#' @return logical TRUE if table was present
+#'
+#' @examples
+#'
+#' if (requireNamespace("RSQLite", quietly = TRUE)) {
+#'   my_db <- dplyr::src_sqlite(":memory:", create = TRUE)
+#'   d <- replyr_copy_to(my_db, data.frame(x=c(1,2)), 'd')
+#'   print(d)
+#'   dplyr::db_list_tables(my_db$con)
+#'   replyr_drop_table_name(my_db, 'd')
+#'   dplyr::db_list_tables(my_db$con)
+#' }
+#'
+#' @export
+#'
+replyr_drop_table_name <- function(dest, name) {
+  if((!is.character(name))||(length(name)!=1)||(nchar(name)<1)) {
+    stop('replyr::replyr_drop_table_name name must be a single non-empty string')
+  }
+  force(dest)
+  if("NULL" %in% class(dest)) {
+    # special "no destination" case
+    return(FALSE)
+  }
+  if('tbl' %in% class(dest)) {
+    # dest was actually another data object, get its source
+    dest <- dest$src
+    if("NULL" %in% class(dest)) {
+      stop("replyr::replyr_drop_table_name unexpected dest")
+    }
+  }
+  # MySQL doesn't seem to always obey overwrite=TRUE
+  # not filing this as MySQL isn't a preferred back end.
+  found = FALSE
+  tryCatch({
+    cn <- dest$con
+    if(!("NULL" %in% class(cn))) {
+      if(name %in% dplyr::db_list_tables(cn)) {
+        found = TRUE
+        dplyr::db_drop_table(cn, name)
+      }
+    }},
+    error=function(x) { warning(x); NULL }
+  )
+  found
+}
 
 
 #' Copy data to remote service.
@@ -21,7 +72,7 @@ NULL
 #'
 #' if (requireNamespace("RSQLite", quietly = TRUE)) {
 #'   my_db <- dplyr::src_sqlite(":memory:", create = TRUE)
-#'   d <- replyr_copy_to(my_db,data.frame(x=c(1,2)),'d')
+#'   d <- replyr_copy_to(my_db, data.frame(x=c(1,2)), 'd')
 #'   print(d)
 #' }
 #'
@@ -31,7 +82,7 @@ replyr_copy_to <- function(dest, df, name = deparse(substitute(df)),
                            rowNumberColumn=NULL) {
   # try to force any errors early, and try to fail prior to side-effects
   if(length(list(...))>0) {
-    stop('replyr_copy_to unexpected arguments')
+    stop('replyr::replyr_copy_to unexpected arguments')
   }
   force(dest)
   if("NULL" %in% class(dest)) {
@@ -48,25 +99,14 @@ replyr_copy_to <- function(dest, df, name = deparse(substitute(df)),
   force(df)
   force(name)
   if("NULL" %in% class(df)) {
-    stop("NULL df to replyr_copy_to")
+    stop("NULL df to replyr::replyr_copy_to")
   }
   if((!is.character(name))||(length(name)!=1)||(nchar(name)<1)) {
-    stop('replyr_copy_to name must be a single non-empty string')
+    stop('replyr::replyr_copy_to name must be a single non-empty string')
   }
-  # MySQL doesn't seem to always obey overwrite=TRUE
-  # not filing this as MySQL isn't a preferred back end.
-  tryCatch({
-    cn <- dest$con
-    if(!("NULL" %in% class(cn))) {
-      if(name %in% dplyr::db_list_tables(cn)) {
-        dplyr::db_drop_table(cn,name)
-      }
-    }},
-    error=function(x) NULL,
-    warning=function(x) NULL
-  )
+  replyr_drop_table_name(dest, name)
   if(!is.null(rowNumberColumn)) {
-    df[[rowNumberColumn]] <- seq_len(nrow(df))
+    df[[rowNumberColumn]] <- seq_len(replyr_nrow(df))
   }
   dplyr::copy_to(dest, df, name,
                  temporary=FALSE,